feat: workspace-awareness for tmux and misc improvements
This commit introduces a significant enhancement to the tmux integration, making it workspace-aware with Aerospace. Now, tmux sessions are automatically created or attached based on the focused Aerospace workspace name. This provides a more seamless and organized workflow.

In addition to the tmux improvements, this commit also includes several other enhancements:

- Brew: The mvn package is now installed by default for all operating systems, and platform-specific package installations have been simplified.

- Neovim:

    - Neovim now automatically changes the current working directory to a directory if it's the only argument passed on startup.

    - A new keymap (<Esc><Esc>) has been added to clear search highlighting.

    - The rust_analyzer LSP server is now installed by default.

- Tmux: The left status bar in tmux is now empty for a cleaner look.

- README: A TODO has been added to dedicate a workspace for ncspot in Aerospace.